requirements.txt
beautifulsoup4
chromadb
fastapi
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
langchain-google-genai
langgraph
lxml
passlib[bcrypt]
pydantic
pydantic-settings
python-magic
python-dotenv
requests
selenium
streamlit
tavily-python
uvicorn[standard]
webdriver-manager
pypdf
sentence-transformers
unstructured
python-multipart
python-jose
pytest
httpx
langmem
langgraph.checkpoint.sqlite